webnfc/layout/index.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/nfchelp.js"></script>
<script>
test(function() {
  assert_throws(null, function() {
    new NFC();
  }, "the constructor should not be callable with 'new'");
  assert_throws(null, function() {
    NFC();
  }, "the constructor should not be callable");
}, "Make sure constructor call type.");

test(function() {
  ['cancelPush', 'push', 'watch', 'cancelWatch']
    .forEach((method) => {
      assert_true(method in navigator.nfc)
    })
}, "Make sure methods are all there.")

test(function() {
  assert_true(navigator.nfc instanceof Object, 'navigator.nfc exists.');
  assert_true('nfc' in navigator, 'navigator.nfc exists.');
}, "Make sure 'navigator.nfc' exists and type is correct.");

//watch
promise_test(t => {
  return navigator.nfc.watch(function() {}).then(result => {
    assert_greater_than(result, 0, 'greater than zero');
  }).catch((err) => {
    assert_reject(err);
  })
}, "Test that nfc watch success if NFC HW is enable.")

promise_test(t => {
 return  promise_rejects(t, "NotSupportedError", navigator.nfc.watch(function() {}));
}, "Test that nfc.watch fails if NFC HW is disable.");

promise_test(t => {
  return promise_rejects(t, "NotSupportedError", navigator.nfc.watch(()=>{}));
}, "Test that nfc.watch fails is NFC HW is not supported.");

//promise_test(t => {
//  return navigator.nfc.watch(noop)
//    .then(() => {
//      assertNFCWatchOptionsEqual(createNFCWatchOptions(), )
//    })
//}, "Test that default NFCWatchOptions values are set correctly.");

promise_test(t => {
  let watchOptions = createNFCWatchOptions(test_message_origin, 'json', 'application/json', 'any');
  return navigator.nfc.watch(noop, watchOptions)
    .then((data) => {
      console.log(data)
    })
},'aaaaaaaaaaaaaaaaaaaaaaaaaaaaa');
</script>
